agent_service/requirements.txt
fastapi
uvicorn
openai
sse-starlette
python-dotenv
